move geocoder input to new file
